requirements.txt
Flask
tensorflow-cpu
numpy
Pillow
scikit-learn
gunicorn